movie_rec/requirements.txt
pandas
numpy
scikit-learn
requests
fastapi
uvicorn
nltk
scipy
python-dotenv
streamlit
imdbpy
#sentence-transformers
sqlalchemy
pytest